Crate pallet_claims

Claims processing pallet

This pallet implement a reward claim mechanism with vesting, for rewarding tokens (CFG or others) awarded through Tinlake investments.

Overview

This pallet is used for processing reward payout claims from investors who invested in Tinlake pools.

Terminology

CFG is the native token of Centrifuge chain.

Usage

Interface

Supported Origins

Valid origin is an administrator or root.

Types

AdminOrigin - Ensure that origin of a transaction is an administrator.

Currency - Expected currency of the reward claim.

Event - Overarching type for pallet events.

MinimalPayoutAmount - Minimal reward payout amount that can be claimed.

PalletId - Constant configuration parameter to store the module identifier for the pallet.

WeightInfo - Pallet’s transaction weights, calculated using runtime benchmarking.

Events

Claimed(T::AccountId, <T as pallet_balances::Config>::Balance) - Event triggered after a reward claim is successfully processed.

RootHashStored(<T as frame_system::Config>::Hash) - Event triggered when the root hash is stored.

Errors

InsufficientBalance - Amount being claimed is less than the available amount in [ClaimedAmounts].

InvalidProofs - The combination of account id, amount, and proofs vector in a claim was invalid.

MustBeAdmin - Protected operation, must be performed by admin.

UnderMinPayout - The payout amount attempting to be claimed is less than the minimum allowed by [MinimalPayoutAmount].

Dispatchable Functions

Callable functions (or extrinsics), also considered as transactions, materialize the pallet contract. Here’s the callable functions implemented in this module:

claim - Claims tokens awarded through tinlake investments.

set_upload_account - Admin function that sets the allowed upload account to add root hashes.

store_root_hash - Stores root hash for correspondent claim merkle tree run.

Public Functions

sorted_hash_of - Build a sorted hash of two given hash values.

Genesis Configuration

The pallet is parameterized and configured via parameter_types macro, at the time the runtime is built by means of the construct_runtime macro.

Dependencies

This pallet is tightly coupled to:

Substrate FRAME’s balances pallet.

References

Substrate FRAME v2 attribute macros.

Credits

The Centrifugians Tribe tribe@centrifuge.io

Re-exports

pub use pallet::*;

pub use crate::traits::WeightInfo as PalletWeightInfo;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

traits