Type Alias pallet_claims::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn claim( origin: OriginFor<T>, account_id: T::AccountId, amount: T::Balance, sorted_hashes: Vec<T::Hash> ) -> DispatchResultWithPostInfo

Claims tokens awarded through Tinlake investments.

Based on hashes length

pub fn set_upload_account( origin: OriginFor<T>, account_id: T::AccountId ) -> DispatchResultWithPostInfo

Admin function that sets the allowed upload account to add root hashes Controlled by custom origin or root

Based on origin check and write op

pub fn store_root_hash( origin: OriginFor<T>, root_hash: T::Hash ) -> DispatchResultWithPostInfo

Stores root hash for correspondent claim Merkle tree run

Based on origin check and write op

impl<T: Config> Pallet<T>

pub fn get_claimed_amount<KArg>(k: KArg) -> T::Balancewhere KArg: EncodeLike<T::AccountId>,

Total claimed amounts for all accounts.

impl<T: Config> Pallet<T>

pub fn get_root_hash() -> Option<T::Hash>

Root hash that correspond to lists of reward claim amounts per account.

impl<T: Config> Pallet<T>

pub fn get_upload_account() -> Option<T::AccountId>

Account that is allowed to upload new root hashes.

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

Return the account identifier of the claims pallet.

This actually does computation. If you need to keep using it, then make sure you cache the value and only call this once.

pub fn sorted_hash_of(a: &T::Hash, b: &T::Hash) -> T::Hash

Build a sorted hash of two given hash values.

Hash a:b if a < b, else b:a. Uses the runtime module’s hasher.